+++ app/src/main/java/com/teamdonut/eatto/ui/home/HomeFragment.java
@@ -1,7 +1,5 @@
 package com.teamdonut.eatto.ui.home;
 
-import android.graphics.Canvas;
-import android.graphics.drawable.Drawable;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -33,7 +31,7 @@ public class HomeFragment extends Fragment {
         initRankRv();
         return binding.getRoot();
     }
-    
+
     void initRankRv(){
         RecyclerView.LayoutManager rankingManager = new LinearLayoutManager(this.getContext()){
             @Override
@@ -46,32 +44,5 @@ public class HomeFragment extends Fragment {
         binding.rvRank.addItemDecoration(new DividerItemDecorator(ContextCompat.getDrawable(getContext(), R.drawable.ranking_divider)));
         binding.rvRank.setHasFixedSize(true);
         binding.rvRank.setLayoutManager(rankingManager);
-    }
-
-    public class DividerItemDecorator extends RecyclerView.ItemDecoration {
-        private Drawable mDivider;
-        public DividerItemDecorator(Drawable divider) {
-            mDivider = divider;
-        }
-
-        @Override
-        public void onDraw(Canvas canvas, RecyclerView parent, RecyclerView.State state) {
-            int space = (int)(parent.getWidth() * 0.03);
-            int dividerLeft = space + parent.getPaddingLeft();
-            int dividerRight = parent.getWidth() - parent.getPaddingRight() - space;
-
-            int childCount = parent.getChildCount();
-            for (int i = 0; i < childCount; i++) {
-                View child = parent.getChildAt(i);
-
-                RecyclerView.LayoutParams params = (RecyclerView.LayoutParams) child.getLayoutParams();
-
-                int dividerTop = child.getBottom() + params.bottomMargin;
-                int dividerBottom = dividerTop + mDivider.getIntrinsicHeight();
-
-                mDivider.setBounds(dividerLeft, dividerTop, dividerRight, dividerBottom);
-                mDivider.draw(canvas);
-            }
-        }
     }
 }
